Added Ultr@VNC's MS-Logon procedure.
----------------------

It appears that...

Ultr@VNC's MS-Logon 1 uses the RFB version string to communicate it's desire to use a different authentication technique.  If the version is 003.004 the Security Negotiation requests it uses VNCAuthentication and does a hybrid of VncAuthentication (DES encrypted with a fixed key) and MS Challenge.

Ultr@VNC's MS-Logon II uses the Security Negotiation protocol (send "-6") to select their MS Logon Authentication procedure.

I have tried to handle both of these.

git-svn-id: http://vncthumbnailviewer.googlecode.com/svn/trunk/VncThumbnailViewer@16 0893bd0a-2c44-0410-b257-bb32fb7553b4